feat: add CategoryService, CategoryController Change CategoryDto's weight to BigDecimal instead of Double

+++ src/main/java/com/example/quality_management_service/form/dto/CategoryDto.java
@@ -2,13 +2,15 @@ package com.example.quality_management_service.form.dto;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CategoryDto(
     Long id,
     @NotNull Long formId,
     @NotNull @Size(max = 30) String title,
-    @NotNull Double weight,
+    @NotNull BigDecimal weight,
     @NotNull Long severityId,
     Integer orderIndex,
     List<FactorDto> factors
